+++ lib/scan/writeBrokenLinks.js
@@ -1,6 +1,6 @@
 const fs = require('fs');
 const path = require('path');
-const checkLink = require("../networking/checkLink.js");
+const checkLink = require("./checkLink.js");
 
 function isLocalhostUrl(url) {
     const localhostRegex = /^(https?:\/\/)?localhost(:\d+)?(\/|$)/i;
@@ -39,17 +39,12 @@ async function writeBrokenLinks(links, PATH, fluff) {
     for (const link of links) {
         if (isLocalhostUrl(link)) { continue }
         try {
-            const result = await new Promise((resolve, reject) => {
-                checkLink(link, function (err, result) {
-                    if (err) {
-                        reject(err);
-                    } else {
-                        resolve(result);
-                    }
-                });
-            });
-            if (result.status === "dead") writeToFile(link, PATH, fluff);
-        } catch {} // This is in a try catch to skip over unsupported urls
+            if (!await checkLink(link)) {
+                writeToFile(link, PATH, fluff);
+            }
+        } catch (error) {
+            console.error('Error:', error);
+        }
     }
     return;
 };
